Fix DeBERTa-v3 FP16 compatibility and hello_modal output
- Remove mixed precision training (GradScaler/autocast): DeBERTa-v3
  stores internal weights in FP16, causing "Attempting to unscale FP16
  gradients" errors with GradScaler
- Add .float() to CEFRModel initialization in train.py, evaluate.py,
  and serve.py to convert all params to FP32, preventing NaN during
  forward passes
- Fix hello_modal.py expected output: platform.node() returns "modal"
  on Modal infrastructure, not a hostname

+++ evaluate.py
@@ -44,7 +44,7 @@ def evaluate():
 
     # Load model
     print("Loading model...")
-    model = CEFRModel()
+    model = CEFRModel().float()
     model.load_state_dict(
         torch.load("/vol/best_model.pt", map_location=device, weights_only=True)
     )
